+++ apps/autotest/models/TST/tst_aclrls.js
@@ -1,1 +1,13 @@
 // eslint-disable-next-line no-undef,camelcase
+const me = tst_aclrls
+
+me.entity.addMethod('testDeleteWithSkip')
+
+/**
+ * Create and return string for fts index
+ * @param {ubMethodParams} ctx
+ * @return {Boolean}
+ */
+me.testDeleteWithSkip = function (ctx) {
+  // 335860161740801
+}